+++ public/vphysics_interfaceV30.h
@@ -882,22 +882,22 @@ public:
 	// parses a text file containing surface prop keys
 	virtual int		ParseSurfaceData( const char *pFilename, const char *pTextfile ) = 0;
 	// current number of entries in the database
-	virtual int		SurfacePropCount( void ) = 0;
+	virtual intp		SurfacePropCount( void ) = 0;
 
-	virtual int		GetSurfaceIndex( const char *pSurfacePropName ) = 0;
-	virtual void	GetPhysicsProperties( int surfaceDataIndex, float *density, float *thickness, float *friction, float *elasticity ) = 0;
+	virtual intp		GetSurfaceIndex( const char *pSurfacePropName ) = 0;
+	virtual void	GetPhysicsProperties( intp surfaceDataIndex, float *density, float *thickness, float *friction, float *elasticity ) = 0;
 
-	virtual surfacedata_t	*GetSurfaceData( int surfaceDataIndex ) = 0;
+	virtual surfacedata_t	*GetSurfaceData( intp surfaceDataIndex ) = 0;
 	virtual const char		*GetString( unsigned short stringTableIndex ) = 0;
 
 
-	virtual const char		*GetPropName( int surfaceDataIndex ) = 0;
+	virtual const char		*GetPropName( intp surfaceDataIndex ) = 0;
 
 	// sets the global index table for world materials
 	virtual void	SetWorldMaterialIndexTable( int *pMapArray, int mapSize ) = 0;
 
 	// NOTE: Same as GetPhysicsProperties, but maybe more convenient
-	virtual void	GetPhysicsParameters( int surfaceDataIndex, surfacephysicsparams_t *pParamsOut ) = 0;
+	virtual void	GetPhysicsParameters( intp surfaceDataIndex, surfacephysicsparams_t *pParamsOut ) = 0;
 };
 
 abstract_class IPhysicsFluidController
